Add joins and exceptions

+++ app/data/db_user.py
@@ -1,9 +1,8 @@
-import email
-import imp
 from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from data.models import DbUser
 from data.hash import Hash
+from fastapi import HTTPException, status
 
 
 def create_user(db: Session, request: UserBase):
@@ -23,15 +22,20 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
-    # TODO: handle exceptions
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
+    return user
 
 # TODO: return user model
 
 
 def update_user(db: Session, id: int, request: UserBase):
-    # TODO: handle exceptions
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
     user.update({
                 DbUser.username: request.username,
                 DbUser.email: request.email,
@@ -43,7 +47,9 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
-# TODO: handle exceptions
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
     db.delete(user)
     db.commit()
     return "ok"
